requirements.txt
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.0.0
spacy==3.7.2
nltk==3.8.1
scikit-learn==1.3.0
gradio==4.7.1
numpy==1.24.0
